+++ src/disk.h
@@ -152,11 +152,13 @@ struct data_index {
 /* Two helper routines for converting sample number (within a major block) and
  * bpm id (as an index into the archive mask) into offsets into a major block
  * for both FA and D data. */
-static inline int fa_data_offset(struct disk_header *header, int sample, int id)
+static inline size_t fa_data_offset(
+    struct disk_header *header, unsigned int sample, unsigned int id)
 {
     return FA_ENTRY_SIZE * (id * header->major_sample_count + sample);
 }
-static inline int d_data_offset(struct disk_header *header, int sample, int id)
+static inline size_t d_data_offset(
+    struct disk_header *header, unsigned int sample, unsigned int id)
 {
     return
         FA_ENTRY_SIZE *
